[unpacking] Fix parsing malformed auto declaration starting with `(`
Change preview check to an assert after last commit.
Handle a tuple declaration that doesn't have `=` after the last `)`. The
`peekPastParen(&token).value == TOK.assign` check is not needed because
the `parseInitializer` parameter is true so `check(TOK.assign)` does the
job and provides a good error message if missing.
Add test for malformed auto and non-auto tuple declarations.